services/professor-impetus/requirements.txt
# Discord
discord.py>=2.3.0

# HTTP client
aiohttp>=3.9.0

# Redis
redis>=5.0.0

# Google Gemini
google-generativeai>=0.3.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
